mtd: nand: brcmnand: Backported fixes from Linux
These are a number of assorted upstream Linux fixes to the
BRCMNAND driver that I have backported in an attempt to get
BRCMBCA working with U-Boot (still not there).

This patch set lowers the hamming distance between the Linux
and U-Boot drivers a bit as well, while we deviate quite
a bit it is still possible to bring fixes over thanks to
exercises like this.

The set tries to prepare the ground for the BCMBCA driver
which I have a port of which is however not yet working
as it should. This is why the read data callback is included.

These patches seem to work fine for me with my devices but
I know the maintainers have some nice test farms so try to
put these to test and see if we can merge them. I bet the
Broadcom folks has this on their TODO list anyway.

To: u-boot@lists.denx.de
To: Dario Binacchi <dario.binacchi@amarulasolutions.com>
To: Michael Trimarchi <michael@amarulasolutions.com>
To: Anand Gore <anand.gore@broadcom.com>
To: William Zhang <william.zhang@broadcom.com>
To: Kursad Oney <kursad.oney@broadcom.com>
To: Philippe Reynes <philippe.reynes@softathome.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

---
Changes in v2:
- Augment the write protect handling to also support the bool
  "write-protect" property in legacy DTS files.
- Fix a mistake in the use_strap DTS property parsing.
- Link to v1: https://patchwork.ozlabs.org/project/uboot/list/?series=423357

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240911-brcmnand-fixes-dabd75230a63",
    "prefixes": [],
    "history": {
      "v1": [
        "20240911-brcmnand-fixes-v1-0-be112a20aaf1@linaro.org"
      ]
    }
  }
}